# server/swagger.yaml
openapi: 3.0.0
info:
  title: Food Sharing App API
  description: API documentation for the Food Sharing App
  version: 1.0.0
servers:
  - url: http://localhost:5020
    description: Local server
paths:
  /api/fridge:
    get:
      summary: Get all fridge items
      responses:
        '200':
          description: A list of fridge items
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    expiryDate:
                      type: string
                      format: date
                    category:
                      type: string
                    available:
                      type: boolean
    post:
      summary: Add a new fridge item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                expiryDate:
                  type: string
                  format: date
                category:
                  type: string
              required:
                - name
                - expiryDate
                - category
      responses:
        '201':
          description: Item added successfully
  /api/fridge/mark-available/{id}:
    post:
      summary: Mark a fridge item as available
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item marked as available
        '404':
          description: Item not found
  /api/fridge/claim/{id}:
    post:
      summary: Claim a fridge item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item claimed successfully
        '404':
          description: Item not found
  /api/groups:
    get:
      summary: Get all groups
      responses:
        '200':
          description: A list of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    tag:
                      type: string
    post:
      summary: Add a new group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                tag:
                  type: string
              required:
                - name
                - tag
      responses:
        '201':
          description: Group added successfully
  /api/alerts:
    get:
      summary: Get all alerts
      responses:
        '200':
          description: A list of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    message:
                      type: string